+++ include/client/RsgClientEngine.hpp
@@ -15,7 +15,7 @@
 #include <thrift/protocol/TMultiplexedProtocol.h>
 
 #include <boost/unordered_map.hpp>
-#include "rsg/RsgServiceImpl.h"
+#include "rsg/services.hpp"
 
 using namespace apache::thrift;
 using namespace apache::thrift::protocol;
