+++ include/rsg/mailbox.hpp
@@ -9,7 +9,7 @@
 #include <boost/unordered_map.hpp>
 #include <boost/shared_ptr.hpp>
 
-#include "rsg/RsgServiceImpl.h"
+#include "rsg/services.hpp"
 #include "actor.hpp"
 #include "comm.hpp"
