+++ include/rsg/services.hpp
@@ -1,0 +1,10 @@
+#ifndef _RSG_SERVICE_
+#define _RSG_SERVICE_
+#include "xbt.h"
+
+#include "services/actorService.hpp"
+#include "services/commService.hpp"
+#include "services/hostService.hpp"
+#include "services/mailboxService.hpp"
+
+#endif /* _RSG_SERVICE_ */
