+++ src/client/RsgClientEngine.cpp
@@ -10,6 +10,8 @@
 #include <string>
 #include <iostream>
 
+using namespace ::simgrid;
+
 XBT_LOG_NEW_CATEGORY(RSG_THRIFT_CLIENT, "Remote SimGrid");
 XBT_LOG_NEW_DEFAULT_SUBCATEGORY(RSG_THRIFT_CLIENT_ENGINE, RSG_THRIFT_CLIENT , "RSG server (Remote SimGrid)");
