+++ src/rsg/services/mailboxService.cpp
@@ -1,0 +1,17 @@
+#include "rsg/services.hpp"
+
+#include "xbt.h"
+#include "simgrid/s4u.h"
+
+#include <iostream>
+
+using namespace ::apache::thrift::server;
+using namespace  ::RsgService;
+
+using namespace ::simgrid;
+
+
+int64_t rsg::RsgMailboxHandler::mb_create(const std::string& name) {
+  s4u::Mailbox *mbox = s4u::Mailbox::byName(name.c_str());
+  return (int64_t) mbox;
+};
